% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ablation.R
\name{csvi_summarize_ablation}
\alias{csvi_summarize_ablation}
\title{Summarize ablation stability}
\usage{
csvi_summarize_ablation(feature_sets, footprint_sets, groups = c("A1", "B1"))
}
\arguments{
\item{feature_sets}{A named list of character vectors or data.frames.}

\item{footprint_sets}{A named list of character vectors or data.frames.}

\item{groups}{Character vector of group names.}
}
\value{
A list of summary tables.
}
\description{
Summarize ablation stability
}
