% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rank.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Jaccard similarity}
\usage{
jaccard_similarity(a, b)
}
\arguments{
\item{a}{Character vector.}

\item{b}{Character vector.}
}
\value{
Numeric scalar.
}
\description{
Jaccard similarity
}
